moving funcs to correct file
